--- src/modules/notes/Note/components/CardNote/style/types/CardNote.style.ts
@@ -1,11 +1,0 @@
-import { Theme } from "@emotion/react";
-import { ViewMode } from "shared/common/types";
-
-export interface ICardNoteContainerStyle {
-	viewmode: ViewMode;
-}
-
-export interface ICardDotStyle {
-	theme?: Theme;
-	color: string;
-}
