+++ src/modules/notes/Note/components/CardNote/types/card-note.types.ts
