+++ src/modules/notes/Note/components/CardNote/types/index.ts
@@ -1,2 +1,2 @@
-export * as default from "./CardNote.types";
-export * from "./CardNote.types";
+export * as default from "./card-note.types";
+export * from "./card-note.types";
