+++ src/modules/notes/Note/components/CardNote/utils/define-grid-size/define-grid-size.ts
