+++ src/modules/notes/Note/components/CardNote/utils/define-grid-size/index.ts
@@ -1,0 +1,1 @@
+export * from "./define-grid-size";
